Projects/Project 01/requirements.txt
grpcio==1.65.1
grpcio-tools==1.65.1
protobuf==4.25.3
